Initial commit = contents of libitzam-6.0.4.tar.gz
The original URL for the Itzam/C library appears to be
unreliable.  (As of today, coyotegulch.com leads to a
generic "parked domain" type page).

Therefore it was required to create a new reliable source
for this open-source software that people depend upon.

See LicenseOpenSource.txt for author, copyright and
license information.